Global Oauth authentication implemented
Signing out of a WordPress.com account clears the OAuth access token.

When a request is made using WPRestClient it now uses the provided
WPRestClient.Authenticator to get an access token. This encapsulates
all of the nitty gritty details of obtaining and saving an access
token inside of WordPress.OauthAuthenticator.