Fixes #437 and other minor issues in the comments list view.
The app doesn't remember selected comments upon resume, or when the "refresh comments" network call ends. In both cases we need to hide the "moderate" bar.

Hide the moderate bar for now sounds ok, but I think the app  should keep a copy of selected comments, and restore them at the end of the refresh/onResume.